Move the files into folders
