+++ three-fps/src/Game/Assets/dirt.jpg
